new features.

+++ UserRegistrationSystem/src/main/java/com/sul/userRegistrationSystem/rest/UserRegistrationRestController.java
@@ -2,6 +2,8 @@ package com.sul.userRegistrationSystem.rest;
 
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,6 +19,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.sul.userRegistrationSystem.dto.CustomErrorType;
 import com.sul.userRegistrationSystem.dto.UserDTO;
 import com.sul.userRegistrationSystem.repository.UserJpaRepository;
 
@@ -30,11 +33,19 @@ public class UserRegistrationRestController {
 	@GetMapping("/list")
 	public ResponseEntity<List<UserDTO>> listAllUsers() {
 		List<UserDTO> users = userJpaRepository.findAll();
+		if(users.isEmpty()) {
+			return new ResponseEntity<List<UserDTO>>(HttpStatus.NO_CONTENT);
+		}
 		return new ResponseEntity<List<UserDTO>>(users, HttpStatus.OK);
 	}
 
 	@PostMapping(value = "/create", consumes = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<UserDTO> createUser(@RequestBody final UserDTO user) {
+	public ResponseEntity<UserDTO> createUser(@Valid @RequestBody final UserDTO user) {
+		if (userJpaRepository.getUserByName(user.getName())!=null) {
+			 return new ResponseEntity<UserDTO>(new CustomErrorType(
+                     "Unable to create new user. A User with name "
+                     + user.getName() + " already exist."),HttpStatus.CONFLICT);
+		}
 		userJpaRepository.save(user);
 		return new ResponseEntity<UserDTO>(user, HttpStatus.CREATED);
 	}
@@ -42,11 +53,20 @@ public class UserRegistrationRestController {
 	@GetMapping("/{id}")
 	public ResponseEntity<UserDTO> getUserById(@PathVariable("id") final Long id) {
 		UserDTO user = userJpaRepository.getUserById(id);
+		if(user==null) {
+			return new ResponseEntity<UserDTO>(new CustomErrorType("User with id " + id +" not found"),HttpStatus.NOT_FOUND);
+		}
 		return new ResponseEntity<UserDTO>(user, HttpStatus.OK);
 	}
 
 	@DeleteMapping("/{id}")
 	public ResponseEntity<UserDTO> deleteUse(@PathVariable("id") final Long id) {
+		UserDTO user = userJpaRepository.getUserById(id);
+        if (user == null) {
+                return new ResponseEntity<UserDTO>(
+                                new CustomErrorType("Unable to delete. User with id "
+                                        + id + " not found."), HttpStatus.NOT_FOUND);
+        }
 		userJpaRepository.deleteById(id);
 		return new ResponseEntity<UserDTO>(HttpStatus.OK);
 	}
@@ -54,6 +74,11 @@ public class UserRegistrationRestController {
 	@PutMapping(value = "/{id}",consumes=MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<UserDTO> putUser(@PathVariable("id") final Long id,@RequestBody final UserDTO user){
 		UserDTO currentUser = userJpaRepository.getUserById(id);
+		if(currentUser==null) {
+			 return new ResponseEntity<UserDTO>(
+                     new CustomErrorType("Unable to upate. User with id "
+                             + id + " not found."), HttpStatus.NOT_FOUND);
+		}
 		String name = user.getName();
 		String email = user.getEmail();
 		String address = user.getAddress();
